Move test files to tests directory
Test files have been reorganized into a dedicated 'tests' directory. This included updating the require paths in each test file to correctly locate the modules being tested. This change will improve project structure and readability.
