+++ tests/array-shape.test.js
@@ -1,4 +1,4 @@
-const arrayShape = require('./schema-array-shape');
+const arrayShape = require('../schema-array-shape');
 
 describe('JSON Schema Converter', () => {
     // ... (previous test cases remain the same)
